:white_check_mark: adda test for getPlugins
