+++ lib/form/scripts.js
@@ -5,8 +5,8 @@ const stream = require('stream');
 const fs = require('fs');
 const path = require('path');
 const _ = require('lodash');
+const utils = require('../util');
 const globalConfig = require('config');
-const exec = require('child_process').execSync;
 const formJS = fs.readdirSync(path.join(__dirname, 'js')).map(script => {
   return fs.readFileSync(path.join(__dirname, 'js', script));
 }).join('\n\n');
@@ -21,39 +21,11 @@ const rendered = (type) => {
     const allSettings = {};
     const allRepeatables = {};
     const inputs = {};
-    let dir = _.get(globalConfig, 'content.plugins.directory', []);
     const ids = {};
     let plugins = type.attributes.map(plugin => {
       return plugin.type;
     });
-    const existing = {};
-    const root = exec('npm root').toString().trim();
-
-    // Check if configuration is set to string
-    if (typeof dir === 'string') {
-      dir = [dir];
-    }
-
-    // Adds path of node_modules to directory
-    if (!Array.isArray(dir) || dir.length === 0) {
-      dir = [root];
-    }
-
-    // Checks if path to node_modules exists
-    else if (dir.find(value => {
-      return value === root;
-    }).length === 0) {
-      dir.push(root);
-    }
-
-    // Adds path of input-plugins to existing
-    dir.forEach((directory) => {
-      fs.readdirSync(directory).forEach((file) => {
-        if (file.indexOf('input-plugin') >= 0 && !existing.hasOwnProperty(file)) {
-          existing[file] = path.join(directory, file);
-        }
-      });
-    });
+    const existing = utils.getPlugins(globalConfig);
 
     plugins = plugins.filter((item, pos, self) => {
       return self.indexOf(item) === pos;
