+++ lib/util.js
@@ -6,6 +6,10 @@
 'use strict';
 
 const util = require('util');
+const exec = require('child_process').execSync;
+const fs = require('fs');
+const path = require('path');
+const _ = require('lodash');
 
 /**
  * Filters an array of objects based on the value of a key in the objects
@@ -46,6 +50,47 @@ exports.splitPop = (input, splitter) => {
   blocks = blocks.join(splitter);
 
   return blocks;
+};
+
+/**
+ * Gets plugins from node_modules and other directories
+ *
+ * @param {object} globalConfig - Config object
+ *
+ * @returns {object} - all input plugins with path
+ */
+exports.getPlugins = (globalConfig) => {
+  const existing = {};
+  const root = exec('npm root').toString().trim();
+  let dir = _.get(globalConfig, 'content.plugins.directory', []);
+
+  // Check if configuration is set to string
+  if (typeof dir === 'string') {
+    dir = [dir];
+  }
+
+  // Adds path of node_modules to directory
+  if (!Array.isArray(dir) || dir.length === 0) {
+    dir = [root];
+  }
+
+  // Checks if path to node_modules exists
+  else if (dir.filter(value => {
+    return value === root;
+  }).length === 0) {
+    dir.push(root);
+  }
+
+  // Adds path of input-plugins to existing
+  dir.forEach((directory) => {
+    fs.readdirSync(directory).forEach((file) => {
+      if (file.indexOf('input-plugin') >= 0 && !existing.hasOwnProperty(file)) {
+        existing[file] = path.join(directory, file);
+      }
+    });
+  });
+
+  return existing;
 };
 
 /**
